backend/shared/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "intel-shared"
version = "0.1.0"
description = "Shared models and utilities for Intel Ingester"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34",
    "botocore>=1.34",
    "pydantic>=2.0",
    "python-ulid>=2.0",
    "tiktoken>=0.6",
    "trafilatura>=1.9",
    "beautifulsoup4>=4.12",
]
